Optimizations.

+++ day_01/src/main.rs
@@ -24,25 +24,25 @@ fn main() {
     result_2, benchmark_2);
 }
 
-fn do_part_1(contents: &str) -> (i64, Duration) {
+fn do_part_1(contents: &str) -> (i32, Duration) {
     let start = Instant::now();
     let mut frequency = 0;
     for line in contents.lines() {
-        frequency += line.parse::<i64>().expect("Error: Not a number.");
+        frequency += line.parse::<i32>().expect("Error: Not a number.");
     }
     let end = Instant::now().duration_since(start);
     return (frequency, end)
 }
 
 /// I used a HashSet to store frequencies as I do not need indexing.
-fn do_part_2(contents: &str) -> (i64, Duration) {
+fn do_part_2(contents: &str) -> (i32, Duration) {
     let start = Instant::now();
     let mut frequency_set = HashSet::new();
     frequency_set.insert(0); // It starts at 0.
     let mut frequency = 0;
     loop {
         for line in contents.lines() {
-            frequency += line.parse::<i64>().expect("Error: Not a number.");
+            frequency += line.parse::<i32>().expect("Error: Not a number.");
             if frequency_set.contains(&frequency) {
                 let end = Instant::now().duration_since(start);
                 return (frequency, end)
